Provide backward compatibility for the dot-local hack
Broker sets both 'Access' and 'AccessLevel' RpcMessage meta
attributes or pass them as is from a parent broker in order
to keep backward compatibility with older brokers that read
'dot-local' in the 'Access' string to expose their '.local'
dir.

The AccessLevel is treated as plain i32, so broker can evaluate
access also for values that are not defined in the enum.

(#14)

+++ src/shvnode.rs
@@ -232,7 +232,7 @@ impl ShvNode {
                 let method = rq.method().unwrap_or_default();
                 for mm in &self.methods {
                     if mm.name == method {
-                        return if rq_access >= mm.access { Some(*mm) } else { None }
+                        return if rq_access >= mm.access as i32 { Some(*mm) } else { None }
                     }
                 }
             }
